Reminded myself why the JSON backend is only half-finished
Turns out that the JSON format doesn't send back the exact same data as the
XML format. For example, the issue endpoint sends back author metadata in
the XML format, whereas the JSON format just returns an author ID, which is
not useful in Redmine 1.0 (the users endpoint was introduced in 1.1).

This time, I put a comment in the code as to why the JSON backend is
half-finished.